pyproject.toml
[tool.poetry]
name = "volcano-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.99.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
mangum = "0.11.0"
python-dotenv = "^1.0.1"
sqlalchemy = "2.0.29"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

cffi = "^1.16.0"
cryptography = "^42.0.7"
virtualenv = "^20.26.2"
bcrypt = {version = "^4.1.3", platform = "manylinux2014_aarch64"}
argon2-cffi = "^23.1.0"
python-multipart = "^0.0.9"
boto3 = "^1.34.125"
dateparser = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
